#914: Moves learningpath testdata to TestData file
